dashboard/requirements.txt
streamlit==1.41.1
psycopg2-binary==2.9.9
pandas==2.2.3
python-dotenv==1.0.0
plotly==6.0.0
